shows summary of scores at the end of doing flashcards
Former-commit-id: 5f004d860f5ea54bc4c881bb3401117a4fc1735e [formerly d6511cb4c74eaa69d5ee655eebd2dc733b1d7227] [formerly 5f004d860f5ea54bc4c881bb3401117a4fc1735e [formerly d6511cb4c74eaa69d5ee655eebd2dc733b1d7227] [formerly b061f48495743fb79d739a1dfaef3ac316db38aa]]
Former-commit-id: 7482cefd5dcb6c47f477aebb76f1b2a3b0bfb3f5 [formerly 4abbc057993d7f6bab16aa96fd5d6977d66efad6] [formerly 65795a524329029dfa7e4c0689fd81cce744e58e] [formerly d9b62a7736f7a614298e1a260f68d3bed6ec63d7 [formerly ca0274ab2566e6a65bb976f2586b38ccdf6be2f3] [formerly 57776a396565354a55653a4b2b01628f783e5290]]
Former-commit-id: d9b62a7736f7a614298e1a260f68d3bed6ec63d7 [formerly ca0274ab2566e6a65bb976f2586b38ccdf6be2f3]
Former-commit-id: 558b3d7aac54958ee882dd42930d701dd10b8dea
Former-commit-id: 7686a639ebcfef287d36af0664b53e53b1809e76

+++ src/mitll/langtest/shared/flashcard/AVPHistoryForList.java
@@ -144,7 +144,7 @@ public class AVPHistoryForList implements IsSerializable {
       return index;
     }
 
-    public String toString() { return "id " + index + " by " + user + " " + getScore() + (isCurrent() ? " current! " : ""); }
+    public String toString() { return "id " + index + " by " + user + " " + ((int)getScore()) + (isCurrent() ? " current! " : ""); }
   }
 
   public String toString() {
